+++ LibVLCSharp.Forms.MediaPlayerElement/ChromecastLocator.cs
@@ -6,16 +6,16 @@ using Xamarin.Forms;
 namespace LibVLCSharp.Forms
 {
     /// <summary>
-    /// Helper class to manage the cast
+    /// Helper class to manage the cast.
     /// </summary>
-    public class ChromecastLocator
+    public static class ChromecastLocator
     {
         /// <summary>
         /// Finds the Chromecast renderers.
         /// </summary>
         /// <param name="libVLC">The <see cref="LibVLC"/> instance.</param>
         /// <returns>An enumerable collection of <see cref="RendererItem"/> representing the Chromecast renderers.</returns>
-        public async Task<IEnumerable<RendererItem>> FindRenderersAsync(LibVLC libVLC)
+        public static async Task<IEnumerable<RendererItem>> FindRenderersAsync(this LibVLC libVLC)
         {
             var renderers = new List<RendererItem>();
             using (var rendererDiscover = new RendererDiscoverer(libVLC,
