+++ LibVLCSharp.Forms.MediaPlayerElement/Converters/ObjectToBoolConverter.cs
@@ -2,7 +2,7 @@
 using System.Globalization;
 using Xamarin.Forms;
 
-namespace LibVLCSharp.Forms
+namespace LibVLCSharp.Forms.Converters
 {
     /// <summary>
     /// Converts not null object to true.
